feat: add plugin events
wch.on() can be used from both the local and remote clients. The first argument is a space-separated list of events. The second argument is the function that will be called with the emitted arguments.

wch.emit() can only be used from the local client (the one available to plugins). The first argument is an event name. The other arguments are sent to all listeners.

wch.off() can be used from both client types. On the remote client, the first and only argument is the listening function. On the local client, the first argument is the space-separated list of events that was passed to wch.on(). The second argument is the listening function.