fix: unwatch only if stream owns its root
The path the stream was created to watch may have an ancestor that is watched.
The stream must avoid unwatching the unowned ancestor (even worse if wch doesn't own it).

The previous roots.has() check didn't account for the `watch` property being consolidated from 2 wch roots,
which always resulted in the ancestor being unwatched.