feat: add offline queue
All methods may be used while the client is not connected.

All actions are always added to the offline queue. If connected, the queue is processed on the next tick of the event loop. Otherwise, it waits until the client reconnects. This makes methods more resilient against disconnects, but disconnects in between the request and response may still cause errors.

Some actions (track, untrack, unsubscribe) will block the queue until they are resolved or rejected. This helps prevent race conditions.